remove new example
