+++ examples/bs-3.1.5/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "reason_language_server.location": "../../lib/bs/native/bin.native"
+}
