+++ examples/bs-3.1.5/bsconfig.json
@@ -1,3 +1,7 @@
+/* This is the BuckleScript configuration file. Note that this is a comment;
+  BuckleScript comes with a JSON parser that supports comments and trailing
+  comma. If this screws with your editor highlighting, please tell us by filing
+  an issue! */
 {
   "name": "react-template",
   "reason": {
@@ -7,7 +11,6 @@
     "dir" : "src",
     "subdirs" : true
   },
-  "ppx-flags": ["@ahrefs/bs-emotion-ppx/bin/bs-emotion-ppx"],
   "package-specs": [{
     "module": "commonjs",
     "in-source": true
