+++ examples/bs-3.1.5/src/Component1.bs.js
@@ -1,0 +1,42 @@
+// Generated by BUCKLESCRIPT VERSION 3.1.5, PLEASE EDIT WITH CARE
+'use strict';
+
+var Curry = require("bs-platform/lib/js/curry.js");
+var React = require("react");
+var ReasonReact = require("reason-react/src/ReasonReact.js");
+
+var component = ReasonReact.statelessComponent("Page");
+
+function handleClick(_, _$1) {
+  console.log("clicked!");
+  return /* () */0;
+}
+
+function make(message, _) {
+  return /* record */[
+          /* debugName */component[/* debugName */0],
+          /* reactClassInternal */component[/* reactClassInternal */1],
+          /* handedOffState */component[/* handedOffState */2],
+          /* willReceiveProps */component[/* willReceiveProps */3],
+          /* didMount */component[/* didMount */4],
+          /* didUpdate */component[/* didUpdate */5],
+          /* willUnmount */component[/* willUnmount */6],
+          /* willUpdate */component[/* willUpdate */7],
+          /* shouldUpdate */component[/* shouldUpdate */8],
+          /* render */(function (self) {
+              return React.createElement("div", {
+                          onClick: Curry._1(self[/* handle */0], handleClick)
+                        }, message);
+            }),
+          /* initialState */component[/* initialState */10],
+          /* retainedProps */component[/* retainedProps */11],
+          /* reducer */component[/* reducer */12],
+          /* subscriptions */component[/* subscriptions */13],
+          /* jsElementWrapped */component[/* jsElementWrapped */14]
+        ];
+}
+
+exports.component = component;
+exports.handleClick = handleClick;
+exports.make = make;
+/* component Not a pure module */
