+++ examples/bs-3.1.5/src/Component1.re
@@ -1,11 +1,6 @@
 /* This is the basic component. */
 let component = ReasonReact.statelessComponent("Page");
 
-let css = _ => ();
-let label = _ => ();
-let prop = _ => ();
-let styles = [%css [prop("1"), prop("2")]];
-
 /* Your familiar handleClick from ReactJS. This mandatorily takes the payload,
    then the `self` record, which contains state (none here), `handle`, `reduce`
    and other utilities */
