+++ examples/bs-3.1.5/src/Index.bs.js
@@ -1,0 +1,13 @@
+// Generated by BUCKLESCRIPT VERSION 3.1.5, PLEASE EDIT WITH CARE
+'use strict';
+
+var ReactDOMRe = require("reason-react/src/ReactDOMRe.js");
+var ReasonReact = require("reason-react/src/ReasonReact.js");
+var Component1$ReactTemplate = require("./Component1.bs.js");
+var Component2$ReactTemplate = require("./Component2.bs.js");
+
+ReactDOMRe.renderToElementWithId(ReasonReact.element(/* None */0, /* None */0, Component1$ReactTemplate.make("Hello!", /* array */[])), "index1");
+
+ReactDOMRe.renderToElementWithId(ReasonReact.element(/* None */0, /* None */0, Component2$ReactTemplate.make("Hello!", /* array */[])), "index2");
+
+/*  Not a pure module */
