--- examples/bs-7.0.1/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "reason_language_server.location": "../../_build/install/default/bin/Bin",
-}
