--- examples/bs-7.0.1/src/Component1.bs.js
@@ -1,60 +1,0 @@
-// Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
-'use strict';
-
-var Curry = require("bs-platform/lib/js/curry.js");
-var React = require("react");
-var ReasonReact = require("reason-react/src/ReasonReact.js");
-
-var component = ReasonReact.statelessComponent("Page");
-
-function css(param) {
-  return /* () */0;
-}
-
-function label(param) {
-  return /* () */0;
-}
-
-function prop(param) {
-  return /* () */0;
-}
-
-function handleClick(_event, _self) {
-  console.log("clicked!");
-  return /* () */0;
-}
-
-function make(message, _children) {
-  return {
-          debugName: component.debugName,
-          reactClassInternal: component.reactClassInternal,
-          handedOffState: component.handedOffState,
-          willReceiveProps: component.willReceiveProps,
-          didMount: component.didMount,
-          didUpdate: component.didUpdate,
-          willUnmount: component.willUnmount,
-          willUpdate: component.willUpdate,
-          shouldUpdate: component.shouldUpdate,
-          render: (function (self) {
-              return React.createElement("div", {
-                          onClick: Curry._1(self.handle, handleClick)
-                        }, message);
-            }),
-          initialState: component.initialState,
-          retainedProps: component.retainedProps,
-          reducer: component.reducer,
-          subscriptions: component.subscriptions,
-          jsElementWrapped: component.jsElementWrapped
-        };
-}
-
-var styles = /* () */0;
-
-exports.component = component;
-exports.css = css;
-exports.label = label;
-exports.prop = prop;
-exports.styles = styles;
-exports.handleClick = handleClick;
-exports.make = make;
-/* component Not a pure module */
