+++ src/analyze_example_tests/ExamplesTests.re
@@ -40,7 +40,7 @@ let projects = [
   ("example-es6-imports", ["src"], "npm install"),
   ("example-react", ["src", "__tests__"], "npm install"),
   ("name_with_underscore", ["src"], "npm install"),
-  ("bs-7.0.1", ["src"], "npm install"),
+  ("bs-3.1.5", ["src"], "npm install"),
   // ("example-esy-dune-project", ["lib", "bin"], "esy"),
 ];
